+++ src/main/java/com/space/server/ai/service/implementation/PromptCreator.java
@@ -3,7 +3,7 @@ package com.space.server.ai.service.implementation;
 import com.space.server.chat.domain.value.Type;
 import com.space.server.domin.chapter.domain.Chapter;
 import com.space.server.core.checklist.domain.Checklist;
-import com.space.server.core.quiz.domain.Quiz;
+import com.space.server.domin.quiz.domain.Quiz;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
