+++ src/main/java/com/space/server/chat/service/CommandChatService.java
@@ -19,8 +19,8 @@ import com.space.server.domin.chapter.domain.Chapter;
 import com.space.server.domin.chapter.service.implementation.ChapterReader;
 import com.space.server.core.checklist.domain.Checklist;
 import com.space.server.core.checklist.service.implementation.ChecklistReader;
-import com.space.server.core.quiz.domain.Quiz;
-import com.space.server.core.quiz.service.implementation.QuizReader;
+import com.space.server.domin.quiz.domain.Quiz;
+import com.space.server.domin.quiz.service.implementation.QuizReader;
 import com.space.server.state.domain.State;
 import com.space.server.state.domain.value.Status;
 import com.space.server.state.service.implementation.StateCreator;
