+++ src/main/java/com/space/server/core/checklist/domain/Checklist.java
@@ -1,7 +1,7 @@
 package com.space.server.core.checklist.domain;
 
 import com.space.server.domin.chapter.domain.Chapter;
-import com.space.server.core.quiz.domain.Quiz;
+import com.space.server.domin.quiz.domain.Quiz;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
