+++ src/main/java/com/space/server/core/checklist/service/QueryChecklistService.java
@@ -2,8 +2,8 @@ package com.space.server.core.checklist.service;
 
 import com.space.server.core.checklist.domain.Checklist;
 import com.space.server.core.checklist.service.implementation.ChecklistReader;
-import com.space.server.core.quiz.domain.Quiz;
-import com.space.server.core.quiz.service.implementation.QuizReader;
+import com.space.server.domin.quiz.domain.Quiz;
+import com.space.server.domin.quiz.service.implementation.QuizReader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
