+++ src/main/java/com/space/server/core/checklist/service/implementation/ChecklistReader.java
@@ -3,7 +3,7 @@ package com.space.server.core.checklist.service.implementation;
 import com.space.server.core.checklist.domain.Checklist;
 import com.space.server.core.checklist.domain.repository.ChecklistRepository;
 import com.space.server.core.checklist.exception.ChecklistNotFoundException;
-import com.space.server.core.quiz.domain.Quiz;
+import com.space.server.domin.quiz.domain.Quiz;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
