+++ src/main/java/com/space/server/domin/quiz/domain/repository/QuizRepository.java
@@ -1,7 +1,7 @@
-package com.space.server.core.quiz.domain.repository;
+package com.space.server.domin.quiz.domain.repository;
 
 import com.space.server.domin.chapter.domain.Chapter;
-import com.space.server.core.quiz.domain.Quiz;
+import com.space.server.domin.quiz.domain.Quiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
