+++ src/main/java/com/space/server/domin/quiz/exception/QuizNotFoundException.java
@@ -1,4 +1,4 @@
-package com.space.server.core.quiz.exception;
+package com.space.server.domin.quiz.exception;
 
 import com.space.server.common.exception.SpaceException;
 import org.springframework.http.HttpStatus;
