+++ src/main/java/com/space/server/domin/quiz/service/CommandQuizService.java
@@ -1,11 +1,11 @@
-package com.space.server.core.quiz.service;
+package com.space.server.domin.quiz.service;
 
 import com.space.server.domin.chapter.service.implementation.ChapterReader;
-import com.space.server.core.quiz.domain.Quiz;
-import com.space.server.core.quiz.service.implementation.QuizCreator;
-import com.space.server.core.quiz.service.implementation.QuizDeleter;
-import com.space.server.core.quiz.service.implementation.QuizReader;
-import com.space.server.core.quiz.service.implementation.QuizUpdater;
+import com.space.server.domin.quiz.domain.Quiz;
+import com.space.server.domin.quiz.service.implementation.QuizCreator;
+import com.space.server.domin.quiz.service.implementation.QuizDeleter;
+import com.space.server.domin.quiz.service.implementation.QuizReader;
+import com.space.server.domin.quiz.service.implementation.QuizUpdater;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
