+++ src/main/java/com/space/server/domin/quiz/service/implementation/QuizCreator.java
@@ -1,7 +1,7 @@
-package com.space.server.core.quiz.service.implementation;
+package com.space.server.domin.quiz.service.implementation;
 
-import com.space.server.core.quiz.domain.Quiz;
-import com.space.server.core.quiz.domain.repository.QuizRepository;
+import com.space.server.domin.quiz.domain.Quiz;
+import com.space.server.domin.quiz.domain.repository.QuizRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
