+++ src/main/java/com/space/server/domin/quiz/service/implementation/QuizReader.java
@@ -1,9 +1,9 @@
-package com.space.server.core.quiz.service.implementation;
+package com.space.server.domin.quiz.service.implementation;
 
 import com.space.server.domin.chapter.domain.Chapter;
-import com.space.server.core.quiz.domain.Quiz;
-import com.space.server.core.quiz.domain.repository.QuizRepository;
-import com.space.server.core.quiz.exception.QuizNotFoundException;
+import com.space.server.domin.quiz.domain.Quiz;
+import com.space.server.domin.quiz.domain.repository.QuizRepository;
+import com.space.server.domin.quiz.exception.QuizNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
