+++ src/main/java/com/space/server/state/domain/repository/StateRepository.java
@@ -1,6 +1,6 @@
 package com.space.server.state.domain.repository;
 
-import com.space.server.core.quiz.domain.Quiz;
+import com.space.server.domin.quiz.domain.Quiz;
 import com.space.server.state.domain.State;
 import com.space.server.user.domain.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
