+++ src/main/java/com/space/server/state/presentation/dto/response/StateResponse.java
@@ -1,6 +1,6 @@
 package com.space.server.state.presentation.dto.response;
 
-import com.space.server.core.quiz.domain.Quiz;
+import com.space.server.domin.quiz.domain.Quiz;
 import com.space.server.state.domain.State;
 import com.space.server.state.domain.value.Status;
 import com.space.server.user.domain.Users;
